pyproject.toml
[project]
name = "django-debug-toolbar-mongo"
description = "MongoDB panel for Django Debug Toolbar"
readme = "README.md"
requires-python = ">=3.12"
version = "2.dev0"

dependencies = [
    "django>=4.2",
    "pymongo>=4.0",
    "django-debug-toolbar>=4.0",
]

classifiers = [
    'Environment :: Web Environment',
    'Framework :: Django',
    'Intended Audience :: Developers',
    'License :: OSI Approved :: MIT License',
    'Operating System :: OS Independent',
    'Programming Language :: Python',
    'Topic :: Database',
    'Topic :: Software Development :: Libraries :: Python Modules',
]

[tool.uv]
package = true

[dependency-groups]
dev = [
    "ruff>=0.7.4",
]

[tool.ruff]
lint.select = ["F", "I", "UP", "DJ", "RUF"]
